RawPatch and AsmPatch methods for ELF interface
These methods allow for patching the in-memory ELF data and can be used
with Save to export the patched ELF to disk. I also added a generic
method for converting virtual addresses to file offsets. This was
required to support using the VA to specify where to place the patch.